Revert "[WIP] React 16+ compliant @nteract/mathjax API"
