+++ packages/mathjax/src/index.js
@@ -2,7 +2,6 @@
 
 import Node from "./node";
 import Context from "./context";
-import Provider from "./provider";
 
-export { Node, Provider, Context };
-export default { Node, Provider, Context };
+export { Node, Context };
+export default { Node, Context };
